modularized the functions in main.js. completely modularized. need to test now.
