+++ src/javascripts/helpers/toggleForm.js
@@ -1,0 +1,10 @@
+const toggleForm = () => {
+  const formDiv = document.querySelector('#form-div');
+  if (formDiv.style.display === 'none') {
+    formDiv.style.display = 'block';
+  } else if (formDiv.style.display === 'block') {
+    formDiv.style.display = 'none';
+  }
+};
+
+export default toggleForm;
